ouster-ros Add dual returns support in client, viz and python
This update reworks the client internals to support more configurable
sensor output, and the dual returns profile available in FW 2.2 in
particular. API compatibility has been preserved as much as possible,
with some exceptions noted in the changelog.

Documentation and support for dual returns in the example ROS driver
is still WIP.